eliminando fs

+++ mybot.js
@@ -32,7 +32,9 @@ const args = message.content.slice(prefix.length).trim().split(' ');
 const command = args.shift().toLocaleLowerCase();
 
 if( command === 'ayuda'){
-    message.channel.send('Lo siento, Este bot aun esta en fase de pruebas!! :pensive: ');
+    message.channel.send('Lo siento, Este bot aun esta en fase BETA!! :pensive: ').then(msg => msg.delete({timeout: 3000}));;
+
+    message.delete();
 }
 
 //Expulsar a un usuario
@@ -42,11 +44,12 @@ if(command === 'kick'){
     
     let perms = message.member.hasPermission("MANAGE_ROLES");
 
+    if(!perms) return message.channel.send("No tienes permisos suficientes, para expulsar personas.");
     if(!mencionado) return message.channel.send('Mencione a un usuario');
     if(!razon) return message.channel.send('Debe escribir una razon para expulsar al usuario\n\n**Ejemplo:** -kick @usuariomencionado El usuario ha roto las reglas del servidor');
 
     message.guild.member(mencionado).kick(razon);
-    message.channel.send(`El usuario :point_right_tone3: **${mencionado.tag}** fue expulsado del servidor.\n\n**Razón: **${razon}`)
+    message.channel.send(`El usuario :point_right_tone3: **${mencionado.tag}** fue expulsado del servidor.\n\n**Razón: **${razon}`);
 
     message.delete();
 
@@ -57,13 +60,13 @@ if(command === 'ban'){
     let mencionado = message.mentions.users.first();
     let razon = args.slice(1).join(' ');
 
-    let perms = message.member.hasPermission("MANAGE_ROLES");
 
+    if(!message.member.hasPermission("BAN_MEMBERS")) return message.channel.send("No tienes permisos suficientes, para banear usuarios.");
     if(!mencionado) return message.channel.send('Mencione a un usuario');
     if(!razon) return message.channel.send('Debe escribir una razon para banear al usuario.\n\n**Ejemplo:** -ban @usuariomencionado El usuario ha roto las reglas del servidor');
 
-    message.guild.member(mencionado).ban(razon);
-    message.channel.send(`El usuario :point_right_tone3: **${mencionado.tag}** fue baneado del servidor.\n\n**Razón: **${razon}`)
+    message.guild.member(mencionado).ban(razon).reason;
+    message.channel.send(`El usuario :point_right_tone3: **${mencionado.tag}** fue baneado del servidor.\n\n**Razón: **${razon}\n\nhttps://giphy.com/gifs/ban-banned-admin-fe4dDMD2cAU5RfEaCU`);
 
     message.delete();
 
@@ -83,7 +86,7 @@ if(command === 'arol'){
     if(!role) return message.channel.send('Rol no encontrado en el servidor.');
 
     miembro.roles.add(role).catch(console.error);
-    message.channel.send(`El rol **${role.name}** fue agregado a **${miembro.user.username}**.`);
+    message.channel.send(`El rol **${role.name}** fue agregado a **<@${miembro.id}>**.`);
 
     message.delete();
 
@@ -103,7 +106,7 @@ if(command === 'erol'){
     if(!role) return message.channel.send('Rol no encontrado en el servidor.');
 
     miembro.roles.remove(role).catch(console.error);
-    message.channel.send(`El rol **${role.name}** fue eliminado de **${miembro.user.username}**.`);
+    message.channel.send(`El rol **${role.name}** fue eliminado de **<@${miembro.id}>**.`);
 
     message.delete();
